requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
Pillow>=8.0.0
tqdm>=4.62.0
scipy>=1.7.0
opencv-python>=4.5.0
einops>=0.4.0
transformers>=4.20.0
diffusers>=0.10.0
clip @ git+https://github.com/openai/CLIP.git
svgwrite>=1.4.0
cairosvg>=2.5.0
